// packages/blog-starter-kit/themes/eplus/generated/graphql.ts
export type AudioBlogFeature = Feature & {
  __typename?: 'AudioBlogFeature';
  /** A flag indicating if the audio blog feature is enabled or not. */
  isEnabled: Scalars['Boolean']['output'];
  /** The voice type for the audio blog. */
  voiceType: AudioBlogVoiceType;
};

/** The voice type for the audio blog. */
export enum AudioBlogVoiceType {
  /** Enum for the female voice type of the audio blog. */
  Female = 'FEMALE',
  /** Enum for the male voice type of the audio blog. */
  Male = 'MALE'
}
// ...
export type Content = {
  __typename?: 'Content';
  /** The HTML version of the content. */
  html: Scalars['String']['output'];
  /** The Markdown version of the content. */
  markdown: Scalars['String']['output'];
  /** The text version from sanitized html content. HTML tags are stripped and only text is returned. */
  text: Scalars['String']['output'];
};
// ...
export type DarkModePreferences = {
  __typename?: 'DarkModePreferences';
  /** A flag indicating if the dark mode is enabled for the publication. */
  enabled?: Maybe<Scalars['Boolean']['output']>;
  /** The custom dark mode logo of the publication. */
  logo?: Maybe<Scalars['String']['output']>;
};
// ...
export type Feature = {
  /** Whether the feature is enabled or not. */
  isEnabled: Scalars['Boolean']['output'];
};
// ...
export type NewsletterFeature = Feature & {
  __typename?: 'NewsletterFeature';
  frequency?: Maybe<NewsletterFrequency>;
  /** A flag indicating if the newsletter feature is enabled or not. */
  isEnabled: Scalars['Boolean']['output'];
};
// ...
export type OpenGraphMetaData = {
  __typename?: 'OpenGraphMetaData';
  /** The image used in og:image tag for SEO purposes. */
  image?: Maybe<Scalars['String']['output']>;
};
// ...
export type PagesPreferences = {
  __typename?: 'PagesPreferences';
  /** A flag indicating if the publication's badge page is enabled. */
  badges?: Maybe<Scalars['Boolean']['output']>;
  /** A flag indicating if the publication's member page is enabled. */
  members?: Maybe<Scalars['Boolean']['output']>;
  /** A flag indicating if the publication's newsletter page is enabled. */
  newsletter?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export type Post = Node & {
// ...
  publication?: Maybe<Publication>;
// ...
};
// ...
export type Preferences = {
  __typename?: 'Preferences';
  /** The publication's darkmode preferences. Can be used to load blog in dark mode by default and add a custom dark mode logo. */
  darkMode?: Maybe<DarkModePreferences>;
  /** A flag indicating if the hashnode's footer branding is disabled for the publication. */
  disableFooterBranding?: Maybe<Scalars['Boolean']['output']>;
  /** An object containing pages enabled for the publication. */
  enabledPages?: Maybe<PagesPreferences>;
  /** A flag indicating if subscription popup needs to be shown to be shown for the publication */
  isSubscriptionModalDisabled?: Maybe<Scalars['Boolean']['output']>;
  /** The selected publication's layout, can be stacked, grid or magazine. */
  layout?: Maybe<PublicationLayout>;
  /** The publication's logo url. */
  logo?: Maybe<Scalars['String']['output']>;
  /** The items in the publication's navigation bar. */
  navbarItems: Array<PublicationNavbarItem>;
};
// ...
export type Publication = Node & {
// ...
  posts: PublicationPostConnection;
// ...
};
// ...
export type PublicationFeatures = {
  __typename?: 'PublicationFeatures';
  /**
   * Audio player for blog posts.
   * @deprecated Audio Blogs are not supported anymore. This field will be removed 18/04/23
   */
  audioBlog: AudioBlogFeature;
  /** Individual styling for the publication. */
  customCSS: CustomCssFeature;
  /** GPT Bot crawler to index the publication. */
  gptBotCrawling: GptBotCrawlingFeature;
  /** Headless CMS for the publication. */
  headlessCMS: HeadlessCmsFeature;
  /** Newsletter feature for the publication which adds a `/newsletter` route for collecting subscribers and allows sending out newsletters. */
  newsletter: NewsletterFeature;
  /** Flag to denote if publication is a pro team's publication. */
  proTeam: ProTeamFeature;
  /** Show the read time for blog posts. */
  readTime: ReadTimeFeature;
  /** Widget that shows up if a text on a blog post is selected. Allows for easy sharing or copying of the selected text. */
  textSelectionSharer: TextSelectionSharerFeature;
  /** Show the view count for blog posts. */
  viewCount: ViewCountFeature;
};
// ...
export enum PublicationLayout {
  /** Changes the layout of blog into grid 3 post cards per row. */
  Grid = 'grid',
  /**
   * Changes the layout of blog into magazine style.
   * This is the newest layout.
   */
  Magazine = 'magazine',
  /** Changes the layout of blog into stacked list of posts. */
  Stacked = 'stacked'
}
// ...
export type ReadTimeFeature = Feature & {
  __typename?: 'ReadTimeFeature';
  /** A flag indicating if the read time feature is enabled or not. */
  isEnabled: Scalars['Boolean']['output'];
};
// ...
export type TextSelectionSharerFeature = Feature & {
  __typename?: 'TextSelectionSharerFeature';
  /** A flag indicating if the text selection sharer feature is enabled or not. */
  isEnabled: Scalars['Boolean']['output'];
};
// ...
export enum UrlPattern {
  /** Post URLs contain the slug (for example `my slug`) and a random id (like `1234`) , e.g. "/my-slug-1234". */
  Default = 'DEFAULT',
  /** Post URLs only contain the slug, e.g. "/my-slug". */
  Simple = 'SIMPLE'
}
// ...
export type ViewCountFeature = Feature & {
  __typename?: 'ViewCountFeature';
  /** A flag indicating if the view count feature is enabled or not. */
  isEnabled: Scalars['Boolean']['output'];
};
// ...
export type PublicationFragment = { __typename?: 'Publication', id: string, imprint?: string | null, title: string, displayTitle?: string | null, url: string, urlPattern: UrlPattern, metaTags?: string | null, favicon?: string | null, isTeam: boolean, headerColor?: string | null, followersCount?: number | null, descriptionSEO?: string | null, features: { __typename?: 'PublicationFeatures', customCSS: { __typename?: 'CustomCSSFeature', isEnabled: boolean, published?: { __typename: 'CustomCSS', home?: string | null, post?: string | null, static?: string | null, homeMinified?: string | null, postMinified?: string | null, staticMinified?: string | null } | null, draft?: { __typename: 'CustomCSS', homeMinified?: string | null, postMinified?: string | null, staticMinified?: string | null } | null }, newsletter: { __typename?: 'NewsletterFeature', isEnabled: boolean }, viewCount: { __typename?: 'ViewCountFeature', isEnabled: boolean }, readTime: { __typename?: 'ReadTimeFeature', isEnabled: boolean }, audioBlog: { __typename?: 'AudioBlogFeature', isEnabled: boolean, voiceType: AudioBlogVoiceType }, textSelectionSharer: { __typename?: 'TextSelectionSharerFeature', isEnabled: boolean } }, about?: { __typename?: 'Content', html: string } | null, author: { __typename?: 'User', id: string, name: string, username: string, profilePicture?: string | null, followersCount: number }, ogMetaData: { __typename?: 'OpenGraphMetaData', image?: string | null }, preferences: { __typename?: 'Preferences', layout?: PublicationLayout | null, logo?: string | null, disableFooterBranding?: boolean | null, enabledPages?: { __typename?: 'PagesPreferences', newsletter?: boolean | null, members?: boolean | null } | null, darkMode?: { __typename?: 'DarkModePreferences', enabled?: boolean | null, logo?: string | null } | null, navbarItems: Array<{ __typename: 'PublicationNavbarItem', id: string, label?: string | null, url?: string | null, priority?: number | null, type: PublicationNavigationType, series?: { __typename: 'Series', id: string } | null, page?: { __typename: 'StaticPage', id: string } | null }> }, links?: { __typename?: 'PublicationLinks', twitter?: string | null, instagram?: string | null, github?: string | null, website?: string | null, hashnode?: string | null, youtube?: string | null, linkedin?: string | null, mastodon?: string | null } | null, integrations?: { __typename?: 'PublicationIntegrations', umamiWebsiteUUID?: string | null, gaTrackingID?: string | null, fbPixelID?: string | null, hotjarSiteID?: string | null, matomoURL?: string | null, matomoSiteID?: string | null, fathomSiteID?: string | null, gTagManagerID?: string | null, fathomCustomDomain?: string | null, fathomCustomDomainEnabled?: boolean | null, plausibleAnalyticsEnabled?: boolean | null, koalaPublicKey?: string | null, msClarityID?: string | null } | null };
// ...
export type TagInitialQuery = { __typename?: 'Query', tag?: { __typename?: 'Tag', id: string, name: string, logo?: string | null, slug: string, tagline?: string | null, followersCount: number, postsCount: number } | null, publication?: { __typename?: 'Publication', id: string, imprint?: string | null, title: string, displayTitle?: string | null, url: string, urlPattern: UrlPattern, metaTags?: string | null, favicon?: string | null, isTeam: boolean, headerColor?: string | null, followersCount?: number | null, descriptionSEO?: string | null, posts: { __typename: 'PublicationPostConnection', edges: Array<{ __typename: 'PostEdge', cursor: string, node: { __typename: 'Post', id: string, title: string, slug: string, publishedAt: string, cuid?: string | null, url: string, subtitle?: string | null, brief: string, readTimeInMinutes: number, views: number, author: { __typename: 'User', id: string, username: string, name: string, profilePicture?: string | null, followersCount: number }, coverImage?: { __typename: 'PostCoverImage', url: string, isPortrait: boolean, isAttributionHidden: boolean } | null } }>, pageInfo: { __typename?: 'PageInfo', endCursor?: string | null, hasNextPage?: boolean | null } }, features: { __typename?: 'PublicationFeatures', customCSS: { __typename?: 'CustomCSSFeature', isEnabled: boolean, published?: { __typename: 'CustomCSS', home?: string | null, post?: string | null, static?: string | null, homeMinified?: string | null, postMinified?: string | null, staticMinified?: string | null } | null, draft?: { __typename: 'CustomCSS', homeMinified?: string | null, postMinified?: string | null, staticMinified?: string | null } | null }, newsletter: { __typename?: 'NewsletterFeature', isEnabled: boolean }, viewCount: { __typename?: 'ViewCountFeature', isEnabled: boolean }, readTime: { __typename?: 'ReadTimeFeature', isEnabled: boolean }, audioBlog: { __typename?: 'AudioBlogFeature', isEnabled: boolean, voiceType: AudioBlogVoiceType }, textSelectionSharer: { __typename?: 'TextSelectionSharerFeature', isEnabled: boolean } }, about?: { __typename?: 'Content', html: string } | null, author: { __typename?: 'User', id: string, name: string, username: string, profilePicture?: string | null, followersCount: number }, ogMetaData: { __typename?: 'OpenGraphMetaData', image?: string | null }, preferences: { __typename?: 'Preferences', layout?: PublicationLayout | null, logo?: string | null, disableFooterBranding?: boolean | null, enabledPages?: { __typename?: 'PagesPreferences', newsletter?: boolean | null, members?: boolean | null } | null, darkMode?: { __typename?: 'DarkModePreferences', enabled?: boolean | null, logo?: string | null } | null, navbarItems: Array<{ __typename: 'PublicationNavbarItem', id: string, label?: string | null, url?: string | null, priority?: number | null, type: PublicationNavigationType, series?: { __typename: 'Series', id: string } | null, page?: { __typename: 'StaticPage', id: string } | null }> }, links?: { __typename?: 'PublicationLinks', twitter?: string | null, instagram?: string | null, github?: string | null, website?: string | null, hashnode?: string | null, youtube?: string | null, linkedin?: string | null, mastodon?: string | null } | null, integrations?: { __typename?: 'PublicationIntegrations', umamiWebsiteUUID?: string | null, gaTrackingID?: string | null, fbPixelID?: string | null, hotjarSiteID?: string | null, matomoURL?: string | null, matomoSiteID?: string | null, fathomSiteID?: string | null, gTagManagerID?: string | null, fathomCustomDomain?: string | null, fathomCustomDomainEnabled?: boolean | null, plausibleAnalyticsEnabled?: boolean | null, koalaPublicKey?: string | null, msClarityID?: string | null } | null } | null };